remove the module 'util'
